Move to core instead of livechat
