+++ apps/meteor/app/livechat/server/startup.js
@@ -12,7 +12,6 @@ import { createDefaultBusinessHourIfNotExists } from './business-hour/Helper';
 import { hasPermission } from '../../authorization/server';
 import { Livechat } from './lib/Livechat';
 import { RoutingManager } from './lib/RoutingManager';
-import { settingsRegenerator } from './settingsRegenerator';
 
 import './roomAccessValidator.internalService';
 
@@ -84,6 +83,4 @@ Meteor.startup(async () => {
 			!user?.roles?.includes('bot') &&
 			Livechat.setUserStatusLivechatIf(user._id, 'not-available', {}, { livechatStatusSystemModified: true }),
 	);
-
-	await settingsRegenerator();
 });
