+++ apps/meteor/server/startup/index.ts
@@ -9,6 +9,7 @@ import './coreApps';
 import './presenceTroubleshoot';
 import '../hooks';
 import '../lib/rooms/roomTypes';
+import '../lib/settingsRegenerator';
 import { isRunningMs } from '../lib/isRunningMs';
 
 // only starts network broker if running in micro services mode
